+++ modules/redos-core/shared/src/test/scala/codes/quine/labo/redos/fuzz/FStringSuite.scala
@@ -7,11 +7,11 @@ import FString._
 class FStringSuite extends munit.FunSuite {
   test("FString.cross") {
     val fs1 = FString(2, IndexedSeq(Wrap('a'), Wrap('b'), Wrap('c')))
-    val fs2 = FString(4, IndexedSeq(Wrap('d'), Repeat(0, None, 1), Wrap('e')))
+    val fs2 = FString(4, IndexedSeq(Wrap('d'), Repeat(0, 1), Wrap('e')))
     assertEquals(
       FString.cross(fs1, fs2, 2, 1),
       (
-        FString(3, IndexedSeq(Wrap('a'), Wrap('b'), Repeat(0, None, 1), Wrap('e'))),
+        FString(3, IndexedSeq(Wrap('a'), Wrap('b'), Repeat(0, 1), Wrap('e'))),
         FString(3, IndexedSeq(Wrap('d'), Wrap('c')))
       )
     )
@@ -22,16 +22,16 @@ class FStringSuite extends munit.FunSuite {
       FString.fix(
         FString(
           2,
-          IndexedSeq(Wrap('a'), Repeat(0, None, 2), Wrap('b'), Repeat(0, Some(2), 1), Wrap('c'), Repeat(0, None, 1))
+          IndexedSeq(Wrap('a'), Repeat(0, 2), Wrap('b'), Repeat(0, 1), Wrap('c'), Repeat(0, 1))
         )
       ),
-      FString(2, IndexedSeq(Wrap('a'), Repeat(0, None, 2), Wrap('b'), Wrap('c')))
+      FString(2, IndexedSeq(Wrap('a'), Repeat(0, 2), Wrap('b'), Wrap('c')))
     )
   }
 
   test("FString#isConstant") {
     assert(FString(2, IndexedSeq(Wrap('a'), Wrap('b'), Wrap('c'))).isConstant)
-    assert(!FString(2, IndexedSeq(Wrap('a'), Repeat(0, None, 1), Wrap('b'))).isConstant)
+    assert(!FString(2, IndexedSeq(Wrap('a'), Repeat(0, 1), Wrap('b'))).isConstant)
   }
 
   test("FString#isEmpty") {
@@ -98,33 +98,29 @@ class FStringSuite extends munit.FunSuite {
 
   test("FString#toUString") {
     assertEquals(
-      FString(2, IndexedSeq(Wrap('a'), Repeat(1, None, 2), Wrap('b'), Wrap('c'), Wrap('d'))).toUString,
+      FString(2, IndexedSeq(Wrap('a'), Repeat(1, 2), Wrap('b'), Wrap('c'), Wrap('d'))).toUString,
       UString.from("abcbcbcd", false)
     )
-    assertEquals(
-      FString(2, IndexedSeq(Wrap('a'), Repeat(1, Some(2), 2), Wrap('b'), Wrap('c'), Wrap('d'))).toUString,
-      UString.from("abcbcd", false)
-    )
     intercept[IllegalArgumentException] {
-      FString(1, IndexedSeq(Repeat(0, None, 2), Wrap('a'), Repeat(0, None, 1))).toUString
+      FString(1, IndexedSeq(Repeat(0, 2), Wrap('a'), Repeat(0, 1))).toUString
     }
   }
 
   test("FString#toString") {
     assertEquals(FString(2, IndexedSeq.empty).toString, "''")
     assertEquals(
-      FString(2, IndexedSeq(Wrap('a'), Repeat(1, None, 2), Wrap('b'), Wrap('c'), Wrap('d'))).toString,
+      FString(2, IndexedSeq(Wrap('a'), Repeat(1, 2), Wrap('b'), Wrap('c'), Wrap('d'))).toString,
       "'a' 'bc'³ 'd'"
     )
     assertEquals(
       FString(
         2,
-        IndexedSeq(Repeat(1, Some(1), 1), Wrap('a'), Repeat(1, None, 1), Wrap('b'), Repeat(0, Some(3), 1), Wrap('c'))
+        IndexedSeq(Repeat(1, 1), Wrap('a'), Repeat(1, 1), Wrap('b'), Repeat(0, 1), Wrap('c'))
       ).toString,
-      "'a' 'b'³ 'c'²⁺¹"
+      "'a'³ 'b'³ 'c'²"
     )
     intercept[IllegalArgumentException] {
-      FString(1, IndexedSeq(Repeat(1, None, 2), Wrap('a'), Repeat(0, None, 1))).toString
+      FString(1, IndexedSeq(Repeat(1, 2), Wrap('a'), Repeat(0, 1))).toString
     }
   }
 }